* MonoDevelop.Ide/MonoDevelop.Ide.FindInFiles/FileProvider.cs: Don't   save the file if nothing has been replaced.
* MonoDevelop.Ide/MonoDevelop.Ide.Templates/SingleFileDescriptionTemplate.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Templates/CodeTranslationFileDescriptionTemplate.cs:
  HashtableToStringArray is not necessary. The string parser service
  can directly handle a string dictionary.

* MonoDevelop.Ide/Makefile.am:
* MonoDevelop.Ide/gtk-gui/gui.stetic:
* MonoDevelop.Ide/gtk-gui/objects.xml:
* MonoDevelop.Ide/MonoDevelop.Ide/Ide.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.csproj:
* MonoDevelop.Ide/MonoDevelop.Ide.Gui/Workbench.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Commands/ToolsCommands.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.ExternalTools/ExternalTool.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.ExternalTools/ExternalToolPanel.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Commands/CustomStringTagProvider.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Gui.Components/StringTagSelectorButton.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Projects.OptionPanels/CustomCommandPanel.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Projects.OptionPanels/CustomCommandWidget.cs:
* MonoDevelop.Ide/gtk-gui/MonoDevelop.Ide.ExternalTools.ExternalToolPanelWidget.cs:
* MonoDevelop.Ide/MonoDevelop.Ide.Projects.OptionPanels/CustomCommandPanelWidget.cs:
* MonoDevelop.Ide/gtk-gui/MonoDevelop.Ide.Gui.Components.StringTagSelectorButton.cs:
* MonoDevelop.Ide/gtk-gui/MonoDevelop.Ide.Projects.OptionPanels.CustomCommandWidget.cs:
  Use the new string tag model. Added a new StringTagSelectorButton
  class.

* MonoDevelop.Ide/MonoDevelop.Ide.Projects.OptionPanels/CombineConfigurationPanel.cs:
  Sort the configurations list.

* MonoDevelop.Core/Makefile.am:
* MonoDevelop.Core/MonoDevelop.Core.csproj:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing:
* MonoDevelop.Core/MonoDevelop.Projects/Project.cs:
* MonoDevelop.Core/MonoDevelop.Projects/Solution.cs:
* MonoDevelop.Core/MonoDevelop.Projects/SolutionItem.cs:
* MonoDevelop.Core/MonoDevelop.Projects/WorkspaceItem.cs:
* MonoDevelop.Core/MonoDevelop.Projects/CustomCommand.cs:
* MonoDevelop.Core/MonoDevelop.Projects/ProjectService.cs:
* MonoDevelop.Core/MonoDevelop.Core/StringParserService.cs:
* MonoDevelop.Core/MonoDevelop.Projects/SolutionEntityItem.cs:
* MonoDevelop.Core/MonoDevelop.Projects/ProjectConfiguration.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/StringTagModel.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/IStringTagModel.cs:
* MonoDevelop.Core/MonoDevelop.Projects/DotNetProjectConfiguration.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/StringTagProvider.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/IStringTagProvider.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/StringTagDescription.cs:
* MonoDevelop.Core/MonoDevelop.Core.StringParsing/StringTagModelDescription.cs:
  Reworked the string parser service. Added an extension point for
  registering tag providers. Added StringTagModel and
  StringTagModelDescription classes.

svn path=/trunk/monodevelop/; revision=154152

+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.FindInFiles/FileProvider.cs
@@ -95,8 +95,11 @@ namespace MonoDevelop.Ide.FindInFiles
 		
 		Document document;
 		StringBuilder buffer = null;
+		bool somethingReplaced;
+		
 		public void BeginReplace ()
 		{
+			somethingReplaced = false;
 			TextReader reader = Open ();
 			buffer = new StringBuilder (reader.ReadToEnd ());
 			reader.Close ();
@@ -111,6 +114,7 @@ namespace MonoDevelop.Ide.FindInFiles
 		
 		public void Replace (int offset, int length, string replacement)
 		{
+			somethingReplaced = true;
 			buffer.Remove (offset, length);
 			buffer.Insert (offset, replacement);
 			if (this.document != null) {
@@ -128,12 +132,12 @@ namespace MonoDevelop.Ide.FindInFiles
 				Gtk.Application.Invoke (delegate { document.TextEditor.EndAtomicUndo (); });
 				return;
 			}
-			if (buffer != null) {
+			if (buffer != null && somethingReplaced) {
 				object attributes = DesktopService.GetFileAttributes (FileName);
 				File.WriteAllText (FileName, buffer.ToString ());
 				DesktopService.SetFileAttributes (FileName, attributes);
-				buffer = null;
 			}
+			buffer = null;
 		}
 	}
 }
